Don't re-copy files for gremlin-server.sh -i
Re-copying files that already exist seems to be a problem for Gremlin Server on windows.  Check of existence first and if present, don't copy. TINKERPOP3-804

+++ gremlin-groovy/src/main/groovy/org/apache/tinkerpop/gremlin/groovy/util/DependencyGrabber.groovy
@@ -96,9 +96,12 @@ class DependencyGrabber {
 
     private static Closure copyTo(final Path path) {
         return { Path p ->
+            // check for existence prior to copying as windows systems seem to have problems with REPLACE_EXISTING
             def copying = path.resolve(p.fileName)
-            Files.copy(p, copying, StandardCopyOption.REPLACE_EXISTING)
-            println "Copying - $copying"
+            if (!copying.toFile().exists()) {
+                Files.copy(p, copying, StandardCopyOption.REPLACE_EXISTING)
+                println "Copying - $copying"
+            }
         }
     }
